Move user auth logics to its own module inside users one
