+++ Apps/Users/OAuth/tests/test_facebook_serializer.py
@@ -3,9 +3,9 @@ from mock import patch
 from mock.mock import MagicMock
 from rest_framework.serializers import ValidationError
 
-from Users.OAuth.serializers import FacebookOAuthSerializer
+from Users.Auth.serializers import UserAuthSerializer
 from Users.models import User
-from Users.serializers import UserAuthSerializer
+from Users.OAuth.serializers import FacebookOAuthSerializer
 
 
 @pytest.mark.django_db
